requirements.txt
torch==2.4.0+cu121
torchvision==0.19.0+cu121
torchaudio==2.4.0+cu121
--extra-index-url https://download.pytorch.org/whl/cu121
pendulum
numpy<2.0
matplotlib
tqdm
monai
pillow
safetensors
opencv-python
pyyaml
gradio